Move postgres related test for getfeatureinfo in appropriate file
